Fix: map index shift; large component split; chained ternary readability
